common: make find_in_stringlist return -1 if not found
We have a similar function in Ear Tag and I noticed this while working on
its code; often it's checked with "if not find_in_model", but this will
falsely fail if the index is 0. Looks like this wouldn't be a problem
in Serial Bowl either way, since we used "is None", but better safe
than sorry.